+++ src/repo/settings/RepoSettingsIndexPage.tsx
@@ -3,6 +3,7 @@ import CheckCircleIcon from 'mdi-react/CheckCircleIcon'
 import prettyBytes from 'pretty-bytes'
 import * as React from 'react'
 import { RouteComponentProps } from 'react-router'
+import { Link } from 'react-router-dom'
 import { Observable, Subject, Subscription } from 'rxjs'
 import { map, switchMap, tap } from 'rxjs/operators'
 import { gql, queryGraphQL } from '../../backend/graphql'
@@ -215,10 +216,9 @@ export class RepoSettingsIndexPage extends React.PureComponent<Props, State> {
                         </>
                     ) : (
                         <div className="alert alert-info">
-                            {/* TODO(keegan): update this to a doc page with instructions */}
-                            This Sourcegraph site doesn't support indexed search. See{' '}
-                            <a href="https://github.com/sourcegraph/sourcegraph/pull/239">PR #239</a> for information on
-                            how to enable it.
+                            This Sourcegraph site has not enabled indexed search. See{' '}
+                            <Link to="/help/admin/search">search documentation</Link> for information on how to enable
+                            it.
                         </div>
                     ))}
             </div>