+++ src/routes.tsx
@@ -4,10 +4,8 @@ import { APIConsole } from './api/APIConsole'
 import { ResetPasswordPage } from './auth/ResetPasswordPage'
 import { SignInPage } from './auth/SignInPage'
 import { SignUpPage } from './auth/SignUpPage'
-import { ErrorNotSupportedPage } from './components/ErrorNotSupportedPage'
 import { DiscussionsPage } from './discussions/DiscussionsPage'
 import { DocSitePage } from './docSite/DocSitePage'
-import { ExplorePage } from './explore/ExplorePage'
 import { ExtensionsArea } from './extensions/ExtensionsArea'
 import { LayoutProps } from './Layout'
 import { SurveyPage } from './marketing/SurveyPage'
@@ -25,7 +23,6 @@ import { SiteInitPage } from './site-admin/SiteInitPage'
 import { RedirectToUserPage } from './user/account/RedirectToUserPage'
 import { RedirectToUserSettings } from './user/account/RedirectToUserSettings'
 import { UserArea } from './user/area/UserArea'
-import { canListAllRepositories } from './util/features'
 
 export interface LayoutRouteComponentProps extends RouteComponentProps<any>, LayoutProps {}
 
@@ -135,11 +132,6 @@ export const routes: ReadonlyArray<LayoutRouteProps> = [
         render: props => <ResetPasswordPage {...props} />,
         exact: true,
         forceNarrowWidth: true,
-    },
-    {
-        path: '/explore',
-        render: props => (canListAllRepositories ? <ExplorePage {...props} /> : <ErrorNotSupportedPage />),
-        exact: true,
     },
     {
         path: '/discussions',